RabbitMqConsumer updated in notification service. Application properties file receives new queue names for credit and debit. Config updated also
