+++ identity-management/identity-management/src/main/java/com/neoclan/identitymanagement/communicationConfig/RabbitMQConfig.java
@@ -28,6 +28,17 @@ public class RabbitMQConfig {
     @Value("${rabbitmq.login.email.routing.key}")
     private String loginEmailRoutingKey;
 
+    @Value("${rabbitmq.credit.email.queue.name}")
+    private String creditEmailQueue;
+
+    @Value("${rabbitmq.debit.email.queue.name}")
+    private String debitEmailQueue;
+
+    @Value("${rabbitmq.credit.email.routing.key}")
+    private String creditEmailRoutingKey;
+
+    @Value("${rabbitmq.debit.email.routing.key}")
+    private String debitEmailRoutingKey;
 
     @Bean
     public TopicExchange notificationExchange() {
@@ -35,10 +46,35 @@ public class RabbitMQConfig {
     }
 
     @Bean
+    public Queue creditEmailQueue() {
+        return new Queue(creditEmailQueue);
+    }
+
+    @Bean
+    public Queue debitEmailQueue() {
+        return new Queue(debitEmailQueue);
+    }
+
+    @Bean
+    public Binding creditEmailBinding() {
+        return BindingBuilder.bind(creditEmailQueue())
+                .to(notificationExchange())
+                .with(creditEmailRoutingKey);
+    }
+
+    @Bean
+    public Binding debitEmailBinding() {
+        return BindingBuilder.bind(debitEmailQueue())
+                .to(notificationExchange())
+                .with(debitEmailRoutingKey);
+    }
+
+    @Bean
     public Queue registrationEmailQueue() {
         return new Queue(registrationEmailQueue);
     }
 
+    @Bean
     public Queue loginEmailQueue() {
         return new Queue(loginEmailQueue);
     }
