+++ identity-management/identity-management/src/main/java/com/neoclan/identitymanagement/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.neoclan.identitymanagement.service;
 
+import com.neoclan.identitymanagement.communicationConfig.RabbitMQProducer;
 import com.neoclan.identitymanagement.dto.Response;
 import com.neoclan.identitymanagement.dto.UserData;
 import com.neoclan.identitymanagement.dto.UserUpdateRequestDto;
@@ -21,6 +22,8 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements UserService {
     private UserRepository userRepository;
     private ModelMapper modelMapper;
+
+    private RabbitMQProducer rabbitMQProducer;
 
     @Override
     public List<Response> fetchAll() {
@@ -78,6 +81,7 @@ public class UserServiceImpl implements UserService {
         emailDetails.setSubject("NeoClan Tech Transaction Alert [Credit : " + userBalanceInfo.getTransactionAmount() + "]");
         emailDetails.setMessage("Credit transaction of " + userBalanceInfo.getTransactionAmount() + " has been performed on your account. Your new account balance is " + user.getAccountBalance());
 
+        rabbitMQProducer.sendCreditEmailNotification(emailDetails); //email notification published via rabbit mq
 
         return Response.builder()
                 .responseCode(ResponseUtils.SUCCESSFUL_TRANSACTION)
@@ -96,12 +100,12 @@ public class UserServiceImpl implements UserService {
         user.setAccountBalance(user.getAccountBalance().subtract(userBalanceInfo.getTransactionAmount()));
         userRepository.save(user);
 
-
         EmailDetails emailDetails = new EmailDetails();
         emailDetails.setRecipient(user.getEmail());
         emailDetails.setSubject("NeoClan Tech Transaction Alert [Debit : " + userBalanceInfo.getTransactionAmount() + "]");
         emailDetails.setMessage("Debit transaction of " + userBalanceInfo.getTransactionAmount() + " has been performed on your account. Your new account balance is " + user.getAccountBalance());
 
+        rabbitMQProducer.sendDebitEmailNotification(emailDetails);//rabbit mq publishes notification to the consumer
 
         return Response.builder()
                 .responseCode(ResponseUtils.SUCCESSFUL_TRANSACTION)
