+++ identity-management/identity-management/src/main/resources/application.properties
@@ -18,8 +18,15 @@ spring.rabbitmq.username=guest
 spring.rabbitmq.password=guest
 
 rabbitmq.neoclan.exchange.name=neoclan_notification_exchange
+
 rabbitmq.registration.email.queue.name=neoclan_registration_email_queue
 rabbitmq.registration.email.routing.key=neoclan_registration_email_routing_key
 
 rabbitmq.login.email.queue.name=neoclan_login_email_queue
 rabbitmq.login.email.routing.key=neoclan_login_email_routing_key
+
+rabbitmq.credit.email.queue.name=neoclan_credit_email_queue
+rabbitmq.credit.email.routing.key=neoclan_credit_email_routing_key
+
+rabbitmq.debit.email.queue.name=neoclan_debit_email_queue
+rabbitmq.debit.email.routing.key=neoclan_debit_email_routing_key
