+++ notification-service/notification-service/src/main/resources/application.properties
@@ -27,3 +27,5 @@ spring.mail.properties.mail.smtp.auth=true
 #rabbitMq configuration
 rabbitmq.registration.email.queue.name=neoclan_registration_email_queue
 rabbitmq.login.email.queue.name=neoclan_login_email_queue
+rabbitmq.credit.email.queue.name=neoclan_credit_email_queue
+rabbitmq.debit.email.queue.name=neoclan_debit_email_queue
