--- transaction-service/transaction-service/src/main/java/com/neoclan/transactionservice/dto/communication/EmailDetails.java
@@ -1,18 +1,0 @@
-package com.neoclan.transactionservice.dto.communication;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@Builder
-public class EmailDetails {
-
-    private String recipient;
-    private String message;
-    private String subject;
-    private String attachment;
-}
