+++ transaction-service/transaction-service/src/main/java/com/neoclan/transactionservice/service/TransactionServiceImpl.java
@@ -1,7 +1,6 @@
 package com.neoclan.transactionservice.service;
 
 import com.neoclan.transactionservice.dto.*;
-import com.neoclan.transactionservice.dto.communication.EmailDetails;
 import com.neoclan.transactionservice.dto.communication.UserBalanceInfo;
 import com.neoclan.transactionservice.dto.communication.UserInfo;
 import com.neoclan.transactionservice.entity.TransactionEntity;
