+++ transaction-service/transaction-service/src/main/resources/application.properties
@@ -10,16 +10,6 @@ server.port=8082
 eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka
 spring.application.name=transaction-service
 
-#rabbitMq configuration
-spring.rabbitmq.host=localhost
-spring.rabbitmq.port=5672
-spring.rabbitmq.username=guest
-spring.rabbitmq.password=guest
 
-rabbitmq.neoclan.exchange.name=neoclan_transaction_exchange
-rabbitmq.credit.email.queue.name=neoclan_credit_email_queue
-rabbitmq.credit.email.routing.key=neoclan_credit_email_routing_key
 
-rabbitmq.debit.email.queue.name=neoclan_debit_email_queue
-rabbitmq.debit.email.routing.key=neoclan_debit_email_routing_key
